+++ bin/miden-cli/src/commands/export.rs
@@ -2,13 +2,11 @@ use std::fs::File;
 use std::io::Write;
 use std::path::PathBuf;
 
+use miden_client::Client;
 use miden_client::account::{Account, AccountFile};
 use miden_client::auth::TransactionAuthenticator;
 use miden_client::store::NoteExportType;
-use miden_client::transaction::AccountInterface;
-use miden_client::utils::Serializable;
-use miden_client::{Client, Word};
-use miden_lib::AuthScheme;
+use miden_client::utils::{Serializable, get_public_keys_from_account};
 use tracing::info;
 
 use crate::errors::CliError;
@@ -159,20 +157,4 @@ async fn export_note<AUTH: TransactionAuthenticator + Sync>(
 
     println!("Successfully exported note {note_id}");
     Ok(file)
-}
-
-/// Gets the public key from the storage of an account. This will only work if the account is
-/// created by the CLI as it expects the account to have the `RpoFalcon512` authentication scheme.
-pub fn get_public_keys_from_account(account: &Account) -> Vec<Word> {
-    let mut pub_keys = vec![];
-    let interface: AccountInterface = account.into();
-
-    for auth in interface.auth() {
-        match auth {
-            AuthScheme::NoAuth => {},
-            AuthScheme::RpoFalcon512 { pub_key } => pub_keys.push(Word::from(*pub_key)),
-        }
-    }
-
-    pub_keys
 }