+++ crates/rust-client/src/utils.rs
@@ -5,7 +5,11 @@ use alloc::string::{String, ToString};
 use alloc::vec::Vec;
 use core::num::ParseIntError;
 
+use miden_lib::AuthScheme;
 use miden_lib::account::faucets::BasicFungibleFaucet;
+use miden_lib::account::interface::AccountInterface;
+use miden_objects::Word;
+use miden_objects::account::Account;
 pub use miden_tx::utils::sync::{LazyLock, RwLock, RwLockReadGuard, RwLockWriteGuard};
 pub use miden_tx::utils::{
     ByteReader,
@@ -101,6 +105,25 @@ pub fn tokens_to_base_units(decimal_str: &str, n_decimals: u8) -> Result<u64, To
 
     // Convert the combined string to an integer
     combined.parse::<u64>().map_err(TokenParseError::ParseU64)
+}
+
+/// Gets the public key from the storage of an account. The function is required to create an
+/// `AccountFile` for exporting accounts in the cli and the web client.
+///
+/// # Arguments
+/// - `account`: The Accounts from which to extract the public keys.
+pub fn get_public_keys_from_account(account: &Account) -> Vec<Word> {
+    let mut pub_keys = vec![];
+    let interface: AccountInterface = account.into();
+
+    for auth in interface.auth() {
+        match auth {
+            AuthScheme::NoAuth => {},
+            AuthScheme::RpoFalcon512 { pub_key } => pub_keys.push(Word::from(*pub_key)),
+        }
+    }
+
+    pub_keys
 }
 
 // TESTS