+++ crates/web-client/test/import_export.test.ts
@@ -2,7 +2,15 @@
 
 import { expect } from "chai";
 import { testingPage } from "./mocha.global.setup.mjs";
-import { clearStore, setupWalletAndFaucet } from "./webClientTestUtils";
+import {
+  clearStore,
+  createNewFaucet,
+  createNewWallet,
+  fundAccountFromFaucet,
+  getAccountBalance,
+  setupWalletAndFaucet,
+  StorageMode,
+} from "./webClientTestUtils";
 
 const exportDb = async () => {
   return await testingPage.evaluate(async () => {
@@ -32,6 +40,23 @@ const getAccount = async (accountId: string) => {
   }, accountId);
 };
 
+const exportAccount = async (accountId: string) => {
+  return await testingPage.evaluate(async (_accountId) => {
+    const client = window.client;
+    const accountId = window.AccountId.fromHex(_accountId);
+    const accountBytes = client.exportAccountFile(accountId);
+    return accountBytes;
+  }, accountId);
+};
+
+const importAccount = async (accountBytes: any) => {
+  return await testingPage.evaluate(async (_accountBytes) => {
+    const client = window.client;
+    await client.importAccountFile(_accountBytes);
+    return;
+  }, accountBytes);
+};
+
 describe("export and import the db", () => {
   it("export db with an account, find the account when re-importing", async () => {
     const { accountCommitment: initialAccountCommitment, accountId } =
@@ -45,5 +70,44 @@ describe("export and import the db", () => {
     const { accountCommitment } = await getAccount(accountId);
 
     expect(accountCommitment).to.equal(initialAccountCommitment);
+  });
+});
+
+describe("export and import account", () => {
+  it("should export and import a private account", async () => {
+    const walletSeed = new Uint8Array(32);
+    crypto.getRandomValues(walletSeed);
+
+    const mutable = false;
+    const storageMode = StorageMode.PRIVATE;
+
+    const initialWallet = await createNewWallet({
+      storageMode,
+      mutable,
+      walletSeed,
+    });
+    const faucet = await createNewFaucet();
+
+    const { targetAccountBalance: initialBalance } =
+      await fundAccountFromFaucet(initialWallet.id, faucet.id);
+    const { accountCommitment: initialCommitment } = await getAccount(
+      initialWallet.id
+    );
+    const exportedAccount = await exportAccount(initialWallet.id);
+    await clearStore();
+
+    await importAccount(exportedAccount);
+
+    const { accountCommitment: restoredCommitment } = await getAccount(
+      initialWallet.id
+    );
+
+    const restoredBalance = await getAccountBalance(
+      initialWallet.id,
+      faucet.id
+    );
+
+    expect(restoredCommitment).to.equal(initialCommitment);
+    expect(restoredBalance.toString()).to.equal(initialBalance);
   });
 });